fix: `findStation` 구현 위치 피드백 반영
