+++ src/main/java/subway/station/service/StationProvider.java
@@ -1,0 +1,7 @@
+package subway.station.service;
+
+import subway.station.repository.domain.Station;
+
+public interface StationProvider {
+    Station findById(Long id);
+}
