+++ src/main/java/subway/station/service/StationService.java
@@ -1,39 +1,14 @@
 package subway.station.service;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import subway.station.service.dto.StationRequest;
 import subway.station.service.dto.StationResponse;
-import subway.station.repository.domain.Station;
-import subway.station.repository.StationRepository;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-@Transactional(readOnly = true)
-public class StationService {
-    private StationRepository stationRepository;
+public interface StationService {
+    StationResponse saveStation(StationRequest stationRequest);
 
-    public StationService(StationRepository stationRepository) {
-        this.stationRepository = stationRepository;
-    }
+    List<StationResponse> findAllStations();
 
-    @Transactional
-    public StationResponse saveStation(StationRequest stationRequest) {
-        Station station = stationRepository.save(new Station(stationRequest.getName()));
-        return StationResponse.from(station);
-    }
-
-    public List<StationResponse> findAllStations() {
-        return stationRepository.findAll().stream()
-                .map(StationResponse::from)
-                .collect(Collectors.toList());
-    }
-
-    @Transactional
-    public void deleteStationById(Long id) {
-        stationRepository.deleteById(id);
-    }
-
+    void deleteStationById(Long id);
 }
